requirements.txt
django==3.0.1
django-crispy-forms==1.8.1
django_cleanup==4.0.0
Pillow==6.2.1
requests==2.20.0
urllib3==1.24.2
certifi==2018.01.18
chardet==3.0.4
idna==2.6
pytz==2019.3